Update vector implementation to use ctor vs init
Like the smart strings, it looks better for vectors
to expose only a pointer to a vector, vs the structure
itself. This way, the only way to interact with a vector
is through the provided functions. This improves
modularity and reduces coupling, if I have it right.

pack.c:
 * update internals to use vector pointers vs structs.
 * update to use size_t instead of unsigned int
   (just makes sense, given the vector length signature)

pack.h:
 * update function headers to match those that use size_t

vector.h:
 * slight update, use ctor and dtor vs init and uninit
 * use size_t instead of unsigned int for vector size
 * move vector structure to vector.c

vector.c:
 * replace unsigned int with size_t appropriately
 * replace init functions with ctors and dtors.