Implement fc_shell_extract and the extraction dtor
In C++, the << and >> operators are insertion and
extraction operators - which is why we're calling
our fc_shell functions that instead of printf and
scanf.

Rationale for extraction_dtor:
Linux provides fopen and fclose to a program for resource allocaiton and
deallocation. It doesn't ask a program to know how to close afile. In
the same way, fc_shell knows how to dispose of an extraction and clean
up the memory - so it provides a function for it vs passing the
responsibility of cleanup to the caller.

fc_shell.h:
 * provide extraction fn and extraction dtor.

fc_shell.c:
 * implement extraction fn and extraction dtor.